grpc.go: change all fmt.Errorf to grpcStatus
grpcStatus implements errors returned by gRPC.  These errors are
serialized and transmitted on the wire between server and client, and allow
for additional data to be transmitted via the Details field in the status
proto.  gRPC service handlers should return an error created by this
package, and gRPC clients should expect a corresponding error to be
returned from the RPC call.

Fixes #81

Signed-off-by: Hui Zhu <teawater@hyper.sh>